requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --allow-unsafe --resolver=backtracking requirements.in
#
absl-py==1.4.0
    # via
    #   -r requirements.in
    #   hcls-imaging-ml-toolkit
    #   tensorboard
    #   tensorflow
astunparse==1.6.3
    # via tensorflow
attrs==23.1.0
    # via hcls-imaging-ml-toolkit
cachetools==5.2.1
    # via
    #   -r requirements.in
    #   google-auth
certifi==2023.5.7
    # via requests
charset-normalizer==3.1.0
    # via requests
flatbuffers==23.5.26
    # via tensorflow
gast==0.4.0
    # via tensorflow
google-api-core[grpc]==2.11.1
    # via
    #   google-api-python-client
    #   google-cloud-pubsub
    #   hcls-imaging-ml-toolkit
google-api-python-client==2.92.0
    # via hcls-imaging-ml-toolkit
google-auth==2.18.1
    # via
    #   -r requirements.in
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
    #   hcls-imaging-ml-toolkit
    #   tensorboard
google-auth-httplib2==0.1.0
    # via
    #   -r requirements.in
    #   google-api-python-client
    #   hcls-imaging-ml-toolkit
google-auth-oauthlib==1.0.0
    # via tensorboard
google-cloud-pubsub==2.17.1
    # via hcls-imaging-ml-toolkit
google-pasta==0.2.0
    # via tensorflow
googleapis-common-protos[grpc]==1.59.1
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.12.6
    # via google-cloud-pubsub
grpcio==1.56.0
    # via
    #   google-api-core
    #   google-cloud-pubsub
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   hcls-imaging-ml-toolkit
    #   tensorboard
    #   tensorflow
    #   tensorflow-serving-api
grpcio-status==1.56.0
    # via
    #   google-api-core
    #   google-cloud-pubsub
h5py==3.9.0
    # via tensorflow
hcls-imaging-ml-toolkit @ git+https://github.com/GoogleCloudPlatform/healthcare.git@57c66830#subdirectory=imaging/ml/toolkit
    # via -r requirements.in
httplib2==0.22.0
    # via
    #   -r requirements.in
    #   google-api-python-client
    #   google-auth-httplib2
    #   hcls-imaging-ml-toolkit
idna==3.4
    # via requests
keras==2.13.1
    # via tensorflow
libclang==16.0.0
    # via tensorflow
markdown==3.4.3
    # via tensorboard
markupsafe==2.1.3
    # via werkzeug
mock==3.0.5
    # via
    #   -r requirements.in
    #   hcls-imaging-ml-toolkit
numpy==1.24.3
    # via
    #   -r requirements.in
    #   h5py
    #   hcls-imaging-ml-toolkit
    #   opencv-python
    #   opt-einsum
    #   tensorboard
    #   tensorflow
oauthlib==3.2.2
    # via requests-oauthlib
opencv-python==4.7.0.72
    # via -r requirements.in
opt-einsum==3.3.0
    # via tensorflow
packaging==23.1
    # via tensorflow
pillow==9.4.0
    # via -r requirements.in
proto-plus==1.22.3
    # via google-cloud-pubsub
protobuf==4.23.4
    # via
    #   google-api-core
    #   google-cloud-pubsub
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
    #   tensorboard
    #   tensorflow
    #   tensorflow-serving-api
psutil==5.9.5
    # via -r requirements.in
pyasn1==0.5.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
pydicom==2.4.1
    # via -r requirements.in
pyparsing==3.1.0
    # via httplib2
requests==2.30.0
    # via
    #   -r requirements.in
    #   google-api-core
    #   requests-mock
    #   requests-oauthlib
    #   requests-toolbelt
    #   tensorboard
requests-mock==1.11.0
    # via -r requirements.in
requests-oauthlib==1.3.1
    # via google-auth-oauthlib
requests-toolbelt==1.0.0
    # via
    #   -r requirements.in
    #   hcls-imaging-ml-toolkit
retrying==1.3.4
    # via hcls-imaging-ml-toolkit
rsa==4.9
    # via google-auth
six==1.16.0
    # via
    #   astunparse
    #   google-auth
    #   google-auth-httplib2
    #   google-pasta
    #   hcls-imaging-ml-toolkit
    #   mock
    #   requests-mock
    #   retrying
    #   tensorflow
tensorboard==2.13.0
    # via tensorflow
tensorboard-data-server==0.7.1
    # via tensorboard
tensorflow==2.13.0
    # via
    #   hcls-imaging-ml-toolkit
    #   tensorflow-serving-api
tensorflow-estimator==2.13.0
    # via tensorflow
tensorflow-io-gcs-filesystem==0.32.0
    # via tensorflow
tensorflow-serving-api==2.12.1
    # via hcls-imaging-ml-toolkit
termcolor==2.3.0
    # via tensorflow
typing==3.7.4.3
    # via hcls-imaging-ml-toolkit
typing-extensions==4.5.0
    # via tensorflow
uritemplate==4.1.1
    # via google-api-python-client
urllib3==1.26.16
    # via
    #   google-auth
    #   hcls-imaging-ml-toolkit
    #   requests
werkzeug==2.3.6
    # via tensorboard
wheel==0.40.0
    # via
    #   astunparse
    #   tensorboard
wrapt==1.15.0
    # via tensorflow

# The following packages are considered to be unsafe in a requirements file:
setuptools==68.0.0
    # via
    #   tensorboard
    #   tensorflow
